requirements.txt
flask
requests
python-binance
openai
apscheduler
pandas
python-dotenv
werkzeug